Generalize UnusedGlobs error to UnusedFixups
Summary: In D84572268, this will be extended to report not just unused globs but also unused buildscript fixups. We want to report as many errors together as possible, so that you do not need to re-run Reindeer multiple times to find out about different categories of errors in the fixup someone is working on.

Reviewed By: diliop

Differential Revision: D84572267

fbshipit-source-id: 934c6a7e594902d4d94dda6f016da6a04de80343