Allow ALL fixups to be version-specific
Summary:
Previously, some fixups like `omit_srcs` were allowed to be platform-specific / version-specific using `cfg`, while other fixups like `omit_targets` were not.

```lang=toml
omit_srcs = ["src/foo/**/*.rs"]  # okay
omit_targets = ["libfoo"]        # okay

['cfg(all(target_os = "linux", version = "^2"))']
omit_srcs = ["src/foo/**/*.rs"]  # okay
omit_targets = ["libfoo"]        # ERROR
```

This makes sense because a `srcs` attribute can expand to a `select` to set a different value by platform, whereas for targets, Reindeer is either going to generate a Buck target for some crate in a package, or not. That cannot be platform-dependent. The existence of a target (at the `buck2 utargets` level) cannot depend on configuration.

However, while targets cannot be platform-dependent, we can certainly allow them to be version-dependent using different `omit_targets` fixups for different versions of a package.

This diff changes the following 7 fixups, which were previously required to apply identically to every version of a package, so that they now are able to be different for different versions.

- `visibility`
- `omit_targets`
- `precise_srcs`
- `python_ext`
- `export_sources`
- `compatible_with`
- `target_compatible_with`

```lang=toml
['cfg(any(version = "<0.1", version = ">=0.3"))']
omit_targets = ["libfoo"]
```

It remains invalid to make any of them platform-dependent. Reindeer will report an error if a fixups.toml uses any of the above 7 fixups in a cfg section that is conditional on something other than `version` and the combinators `any`/`all`/`not`.

Reviewed By: diliop

Differential Revision: D84881746

fbshipit-source-id: 3135cac7b7f163b14fa1fa5f9f66d61adbbf9c4f