Enable strict glob checking within all used fixups
Summary:
The `strict_globs` setting in reindeer.toml was introduced 4 years ago in {D32769278}. At the time, it was considered likely too noisy to enable by default. Fixups and the mechanisms for making them conditional were not expressive enough for us to be confident that a fixup someone needed could always be expressed in a way that didn't have some globs being considered unused.

Since then the following work has improved how globs are evaluated, solving false positives *and* false negatives in the strict glob checking.

- The new glob checker from the stack containing D78222867 and D78222868 and D78222866

- The new approach to evaluating conditional fixups unifying platform-specific and version-specific conditionals in D78248586 and D76480482

- Allowing the `export_sources` fixup (which involves globs) to vary by version in D84881746

- Tracking which cfg predicates have matched some combination of package version and platform in D84572268

At this point I am confident that treating globs that Reindeer considers unused as always an error is fine.

However, we will suppress unmatched globs errors inside conditional fixups that have not applied to any package version in any platform. For example:

```lang=toml
['cfg(target_os = "android")']
# checked, but only if you have at least one Android platform
extra_srcs = ["*.asdf"]

['cfg(version = "^3")']
# checked, but only if 3.x of this package is in your dependency graph
extra_srcs = ["*.asdf"]
```

Doing it this way allows a fixups repository like <https://github.com/gilescope/buck2-fixups> to supply ready-to-use fixups for a broad range of platforms and crate versions. The user should not be required to edit out platforms and versions they are not currently using when copying one of these fixups.

We may wish to introduce a separate setting that opts in to reporting entire conditional fixups as unused (the entire `cfg(target_os = "android")` or `cfg(version = "^3")` section). That makes more sense than reporting individual globs (if any) inside that section as unmatched.

Reviewed By: diliop

Differential Revision: D84881747

fbshipit-source-id: 53a2dd87cd387b3da01785b17da1e11d717843f9