Report visibility fixups that do not apply to a public package version
Summary:
The checking for `visibility` fixups was previously only considering the package name, not the package version. A `visibility` fixup was allowed for any package that your workspace directly depends on, and disallowed for transitive-only dependencies (since they are always produced with private visibility).

This had false negatives in the presence of version-specific fixups. For example if we have a workspace that depends on `thiserror = "2"` and this fixup for thiserror:

```lang=toml,name=fixups/thiserror/fixups.toml
['cfg(version = "1")']
visibility = ...
```

Previously this was not reported as unused. After this diff it will be reported as unused because only thiserror v2 would be a public package, not thiserror v1.

Reviewed By: diliop

Differential Revision: D85524909

fbshipit-source-id: 16348b3b3b84ef48a26bd6d33f75be582f010cfa