Support cfgs=['key="value"']
Summary:
This is necessary for correctly interpreting the following case. In `py-polars/fixups.toml` we set `cfgs = ['allocator="mimalloc"']`:

https://www.internalfb.com/code/fbsource/[3b9198910fd1794751fc11b8ea94bbb4f69d5568]/third-party/rust/fixups/py-polars/fixups.toml?lines=18-20

The `py-polars` crate contains a dependency on `either` conditional on this cfg: `cfg(all(target_family = "unix", target_os = "macos", not(allocator = "mimalloc"), not(allocator = "default")))`

https://www.internalfb.com/code/fbsource/[3b9198910fd1794751fc11b8ea94bbb4f69d5568]/third-party/rust/vendor/py-polars-1.29.0/Cargo.toml?lines=145-146

When `allocator = "mimalloc"`, `py-polars` is not supposed to depend on `either`. Previously cfgs fixups in the form `key="value"` did not correctly affect dependency selection in this way.

Reviewed By: diliop

Differential Revision: D85514861

fbshipit-source-id: 73fb4be0f8cbdb202a2554c4edc89bd4c9863695