Add buildscript-specific extra_srcs and omit_srcs
Summary:
Previously, the fixups `extra_srcs` and `extra_mapped_srcs` would apply to all crates in the package, including the build script.

For example we have this `extra_srcs = ["src/proto/specs/*.proto"]` fixup for the `plexi-core` crate:

https://www.internalfb.com/code/fbsource/[c0adcbc455202727140b948b6d25df5d8d8e67d8]/third-party/rust/fixups/plexi-core/fixups.toml?lines=1

That currently adds to the sources of both the plexi-core library and the plexi-core build script.

https://www.internalfb.com/code/fbsource/[c0adcbc455202727140b948b6d25df5d8d8e67d8]/third-party/rust/BUCK?lines=166881-166883%2C166891-166892%2C166937%2C166939-166941%2C166943-166944%2C166959

In reality, only the build script needs this extra source. The library does not.

Inversely, for `rerun`, we have `extra_srcs = ["Cargo.toml"]`:

https://www.internalfb.com/code/fbsource/[ea100f737d0f7500a3e645b167569849abfef378]/third-party/rust/fixups/rerun/fixups.toml?lines=2

but this time only the library compilation needs this source (through `document-features`), not the build script compilation.

https://www.internalfb.com/code/fbsource/[ea100f737d0f7500a3e645b167569849abfef378]/third-party/rust/vendor/rerun-0.22.1/src/lib.rs?lines=12

Going forward, extra sources for build scripts can be specified through the `buildscript.build` section in fixups.toml, which is already used for build script compilation environment variables (since {D73977930}), deps (since {D74158951}), and link style (since {D74046797}).

```lang=toml
extra_srcs = [...]

[buildscript.build]
extra_srcs = [...]
```

`omit_srcs` will continue to apply to all crates in the package. The main use case for `omit_srcs` is files that are not going to be checked into source control, and I don't know of any reason to configure it separately for a library and build script.

Reviewed By: diliop

Differential Revision: D86391704

fbshipit-source-id: df43badc2b5f5d7ef44d241bcadffb7124d027b6