allow `preferred_linkage` in prebuilt_cxx_library fixups (#82)
Summary:
sometimes (e.g. when compiling wasm and `wit-bindgen-rt`) you need to be able to specify that a particular prebuilt library should be used as a static library, and not converted to a shared library.

the best way to force this seems to be `preferred_linkage` on the generated `prebuilt_cxx_library` itself.

Pull Request resolved: https://github.com/facebookincubator/reindeer/pull/82

Reviewed By: cjlongoria

Differential Revision: D86579834

Pulled By: dtolnay

fbshipit-source-id: d4640a858b9ec01af2055a9d8267fe84dc5a5144