Clean up srcfiles tests
Summary:
This incantation used to be repeated in every test:

https://www.internalfb.com/code/fbsource/[e3acb64787c0fba7fae1f11369b789de8c03811e]/fbcode/common/rust/tools/reindeer/src/srcfiles.rs?lines=627-633

It is not obvious at first why this isn't just `assert!(res.errors.is_empty())`, but it's because that would not render the errors into the panic message if the test were to fail. Replaced this with an explicit panic that is more obvious.

Extracted repeated comparison logic from each test to a `assert_srcfiles` function.

`res.files` used to be compared using `HashSet`, but that produces annoying test output because the left-hand side and the right-hand side will each be scrambled in a different random order when the test fails. Replaced with `BTreeSet` which will output according to `Path`'s ordering.

Removed the requirement for commas in the macro call.

Removed weird use of `stringify`.

Reviewed By: diliop

Differential Revision: D86599779

fbshipit-source-id: c7eb0a93107a6e7b03f69830b6ef8ff7345f0cd0