Support artifact dependencies on a specific bin
Summary:
I wanted to use this Cargo functionality in {D86376758}.

```lang=toml
[dependencies]
jj-cli = { version = "0.35", artifact = "bin:jj" }
```

Cargo allows you to specify a subset of the dependency's binaries you want to depend on. Previously, Reindeer only supported `artifact = "bin"` which depends on *all* of a package's binaries, in this case including `bin:jj` and `bin:fake-bisector` and `bin:fake-diff-editor` and `bin:fake-editor` and `bin:fake-formatter`. This diff implements depending on a specific binary.

Reviewed By: diliop

Differential Revision: D87012064

fbshipit-source-id: f1a7ac28c35ffdcbfc2251c30bee217d71663c37