Remove rename from private rule name
Summary:
If we have a top-level Cargo.toml containing `syn = "2"` and `syn1 = { package = "syn", version = "1" }`, Reindeer produces public targets for these named `syn` and `syn1` (obviously) and produces private targets `syn-2.0.110` and `syn-1.0.109-syn1`.

https://www.internalfb.com/code/fbsource/[54ea9b8e13ee36f031559d2b35a0548c1f9502ec]/third-party/rust/BUCK?lines=261186-261188%2C261218%2C261316-261318%2C261348

I was not able to figure out any reason that the rename "syn1" would need to appear in the name of the private target. It appears to be an accident of history around the time of {D40591188}.

This commit renames the generated private target from `syn-1.0.109-syn1` to just `syn-1.0.109`. This will make the upcoming split third-party/rust easier to do by having target names that are only a function of the package name and version.

Reviewed By: diliop

Differential Revision: D87690032

fbshipit-source-id: 0e4f962adbfc24298c43bd60ee969ff237264add