Remove Index::private_rule_name
Summary:
After {D87690032}, `index.private_rule_name(pkg)` consists only of `Name(pkg.to_string())`, making no use of `index`.

https://www.internalfb.com/code/fbsource/[6cc689734356cc5744a2578c1ac6028c3be69d41]/fbcode/common/rust/tools/reindeer/src/index.rs?lines=85%2C239-241%2C359

As part of splitting the third-party/rust package, `private_rule_name` is an obstacle because rules are no longer going to be identified only by a target name like `syn-2.0.110`, but also by the package containing that target: `//third-party/rust/vendor/syn:2.0.110`. But some fraction of the callers of `private_rule_name` want to use the resulting name as a dependency (for which having the package path included is very important):

https://www.internalfb.com/code/fbsource/[6cc689734356cc5744a2578c1ac6028c3be69d41]/fbcode/common/rust/tools/reindeer/src/fixups.rs?lines=808%2C813-814%2C911%2C913%2C922%2C925-926

whereas other callers want to use the resulting name as a fragment for constructing the name of some other local target in the same package (for which the package path needs to not be included):

https://www.internalfb.com/code/fbsource/[6cc689734356cc5744a2578c1ac6028c3be69d41]/fbcode/common/rust/tools/reindeer/src/buckify.rs?lines=893-894%2C898-899%2C940-941%2C944-946

This commit inlines the logic of `private_rule_name` to all the call sites, and each one will be able to be adjusted separately as needed for the third-party/rust split.

Reviewed By: diliop

Differential Revision: D87690030

fbshipit-source-id: 17407387dc6f4af47e121e550f92726143af5fb3