Generalize Alias.actual to non-local targets
Summary:
For splitting third-party/rust, we will need aliases that are able to refer to targets in a different package. For example `//third-party/rust:syn` will be an alias to `//third-party/rust/vendor/syn:2.0.110`.

So far, `Alias` has encoded an assumption that aliases only refer to targets within the same package. This commit generalizes that.

Reviewed By: diliop

Differential Revision: D87690031

fbshipit-source-id: 0db27b77aeac16ad02ef3cddfc08ff4f99f99a8e