Disable `licenses` in split mode
Summary:
The `licenses` attribute of rust_library and rust_binary contains paths to license files as `attrs.list(attrs.source())`.

https://www.internalfb.com/code/fbsource/[a2f00c570f199d6349b101bbbd25e267e3125b8c]/third-party/rust/BUCK?lines=3740-3741%2C3781-3785%2C3833

In split third-party/rust, this doesn't work in a straightforward way because a `rust_library` target under `third-party/rust/vendor/aho-corasick/BUCK` (for example) is not allowed to refer to a source located in a different Buck package, `"../aho-corasick-1.1.3/COPYING"`.

There are ways we can make this work with some effort, using `export_file` or building it into the `RustSources` provider from {D87691396}. But according to igorsugak, we currently have nothing that looks at `licenses`. All the analysis we do with licenses in the build graph is done using METADATA.bzl and "third-party.metadata" ({D87389748}), both for OSS compliance testing and for SBOM generation in fbpkg. So for now there is no need to block split third-party/rust on solving `licenses`.

Reviewed By: capickett

Differential Revision: D87883872

fbshipit-source-id: b8e86526495f9d271ef5e47aded8053318068dcf