Compute package path by finding buck root
Summary:
Unlike when generating a single BUCK file, in split third-party/rust mode the output BUCK files will need to contain different content depending on where in the repository they are located. For example if the third-party directory is `//third-party/rust`, we might have visibilities that look like `visibility = ["//third-party/rust/..."]` and deps that look like `deps = ["//third-party/rust/vendor/quote:1.0.100"]`.

If it turns out to make sense in the future, I guess it might be possible to generate location-independent output like `deps = ["//" + native.package_name().removesuffix("/syn") + "/quote:1.0.100"]` but it seems silly to go in that direction prematurely. For now we can just find the location of the repository root based on the presence of `.buckconfig`/`.buckroot` and write that hardcoded package path into the generated files.

Reviewed By: diliop

Differential Revision: D87949539

fbshipit-source-id: f3a3cde5850c32ad17db5179937cfe895ab76207