Combine BuildscriptGenrule manifest dir fields into enum
Summary: In {D88033642}, I will need to expand BuildscriptGenrule with a 4th possibility for how a manifest directory is provided to the build script execution. In addition to being None or BuckPath or Subtarget, it will also need to be RuleRef. All of these possibilities are mutually exclusive and better represented using an enum than individual Option fields.

Reviewed By: diliop

Differential Revision: D88033637

fbshipit-source-id: f7a11a28d2b10f01274d729cb3fc72b1df79a5b1