Relocate rust-third-party root package to subdirectory
Summary:
The `name = "rust-third-party"` top-level target is just one of the 1494 targets in //third-party/rust, but accounts for over 80% of the remaining load time of the package (after splitting out everything else in this stack) as measured by `buck2 profile loading fbsource//third-party/rust:`.

This diff relocates it to `fbsource//third-party/rust/top:rust-third-party` with a cheaply loadable alias in the original location.

Reviewed By: diliop

Differential Revision: D88039664

fbshipit-source-id: 03fe2c5ecd93a83aeb778aa27a8ad69bd7750214