Erase vendored BUCK files in crates that are not buckified
Summary:
D88102551 failed to land because of 3 files present in the result of `reindeer vendor` that were not supposed to be:

- third-party/rust/vendor/lock_free_hashtable-0.1.1/BUCK
- third-party/rust/vendor/sorted_vector_map-0.2.0/BUCK
- third-party/rust/vendor/static_interner-0.1.1/BUCK

These are crates that are vendored by `cargo vendor` but **not** buckified by `reindeer buckify`, either because it is a platform-specific dependency for some platform that Reindeer is not configured for, or because of an omit_deps fixup.

Because third-party/rust/.gitignore no longer encompasses `third-party/rust/vendor/*/BUCK` since D88033647, we need `reindeer vendor` to delete these files.

Reviewed By: capickett

Differential Revision: D88107990

fbshipit-source-id: 0c9d4e981f2e703b50a80c3b707522dcaaaaafbb